% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AME_fit_collapse_gash.R
\name{AME.collapse.crossfit.boot}
\alias{AME.collapse.crossfit.boot}
\title{Estimating PAMCE withithout regularization}
\usage{
AME.collapse.crossfit.boot(formula, data, pair = FALSE, fac.level,
  ord.fac, nway = 1, cv.collapse.cost = c(0.1, 0.3, 0.5, 0.7),
  cv.type = "cv.1Std", marginal_dist, marginal_type, family,
  difference = FALSE, boot = 100, tableAME_base, coefAME_base_l,
  numCores, seed)
}
\arguments{
\item{formula}{formula}

\item{data}{data}

\item{pair}{whether we use the paired-conjoint design}

\item{pair_id}{id for paired-conjoint design. required when 'pair = TRUE'}
}
\description{
Estimating PAMCE withithout regularization
}
