% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AME_fit_basic.R
\name{AME.fit}
\alias{AME.fit}
\title{Estimating PAMCE withithout regularization}
\usage{
AME.fit(formula_full, data, pair = FALSE, marginal_dist,
  marginal_dist_u_list, marginal_dist_u_base, marginal_type,
  difference = FALSE)
}
\arguments{
\item{data}{data}

\item{pair}{whether we use the paired-conjoint design}

\item{formula}{formula}

\item{pair_id}{id for paired-conjoint design. required when 'pair = TRUE'}
}
\description{
Estimating PAMCE withithout regularization
}
