% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AME_estimate_collapse_gash.R
\name{AME_estimate_collapse_gash}
\alias{AME_estimate_collapse_gash}
\title{Estimating PAMCE withithout regularization}
\usage{
AME_estimate_collapse_gash(formula, data, ord.fac, pair = FALSE,
  pair_id = NULL, cluster = NULL, marginal_dist, marginal_type,
  difference = FALSE, family = "binomial", nway = 1,
  cv.collapse.cost = c(0.1, 0.3, 0.5, 0.7), cv.type = "cv.1Std",
  boot = 100, numCores, seed)
}
\arguments{
\item{formula}{formula}

\item{data}{data}

\item{pair}{whether we use the paired-conjoint design}

\item{pair_id}{id for paired-conjoint design. required when 'pair = TRUE'}
}
\description{
Estimating PAMCE withithout regularization
}
