% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AME_estimate_full.R
\name{AME_estimate_full}
\alias{AME_estimate_full}
\title{Estimating the population AMCE using the model-based approach}
\usage{
AME_estimate_full(formula, formula_three = NULL, data,
  type = "genlasso", ord.fac, pair = FALSE, pair_id = NULL,
  cross_int = TRUE, cluster = NULL, target_dist, target_type,
  difference = FALSE, cv.type = "cv.1Std", nfolds = 5, boot = 100,
  seed = 1234, numCores = NULL)
}
\arguments{
\item{formula}{Formula}

\item{formula_three}{Formula for three-way interactions}

\item{data}{Data}

\item{type}{"No-Reg" or "genlasso"}

\item{ord.fac}{Whether we assume each factor is ordered. When not specified, we assume all of them are ordered}

\item{pair}{Whether we use a paired-choice conjoint design}

\item{cross_int}{Include interactions across profiles}

\item{cluster}{unique identifiers for cluster}

\item{difference}{Whether we compute the difference between different pAMCEs}

\item{cv.type}{(when type =genlasso)  `cv.1Std`` (stronger) or `cv.min` (weaker).}

\item{boot}{The number of bootstrap samples}

\item{seed}{Seed for bootstrap}

\item{marginal_dist}{Marginal distributions of profiles to be used}

\item{marginal_type}{Names of marginal distributions}
}
\description{
Estimating the population AMCE using the model-based approach
}
