% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AME_fit_basic.R
\name{cAME_from_boot}
\alias{cAME_from_boot}
\title{Estimating PAMCE withithout regularization}
\usage{
cAME_from_boot(outAME, factor_name, level_name, difference = TRUE)
}
\arguments{
\item{formula}{formula}

\item{data}{data}

\item{pair}{whether we use the paired-conjoint design}

\item{pair_id}{id for paired-conjoint design. required when 'pair = TRUE'}
}
\description{
Estimating PAMCE withithout regularization
}
