% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AME_fit_collapse_gash.R
\name{collapse.fit}
\alias{collapse.fit}
\title{Estimating PAMCE withithout regularization}
\usage{
collapse.fit(formula, family, data, pair = TRUE, nway = 1,
  collapse = TRUE, collapse.cost, fac.level, ord.fac)
}
\arguments{
\item{formula}{formula}

\item{data}{data}

\item{pair}{whether we use the paired-conjoint design}

\item{pair_id}{id for paired-conjoint design. required when 'pair = TRUE'}
}
\description{
Estimating PAMCE withithout regularization
}
