% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Collapsing.R
\name{Collapsing}
\alias{Collapsing}
\title{Estimating PAMCE withithout regularization}
\usage{
Collapsing(fit)
}
\arguments{
\item{formula}{formula}

\item{data}{data}

\item{pair}{whether we use the paired-conjoint design}

\item{pair_id}{id for paired-conjoint design. required when 'pair = TRUE'}
}
\description{
Estimating PAMCE withithout regularization
}
