% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AME_fit_basic.R
\name{createDist}
\alias{createDist}
\title{Estimating PAMCE withithout regularization}
\usage{
createDist(formula, data)
}
\arguments{
\item{formula}{formula}

\item{data}{data}

\item{pair}{whether we use the paired-conjoint design}

\item{pair_id}{id for paired-conjoint design. required when 'pair = TRUE'}
}
\description{
Estimating PAMCE withithout regularization
}
