% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreatelevelIndex.R
\name{CreatelevelIndex}
\alias{CreatelevelIndex}
\title{Estimating PAMCE withithout regularization}
\usage{
CreatelevelIndex(fac.level, ord.fac, indTwo = NULL, indThree = NULL,
  Gorder)
}
\arguments{
\item{formula}{formula}

\item{data}{data}

\item{pair}{whether we use the paired-conjoint design}

\item{pair_id}{id for paired-conjoint design. required when 'pair = TRUE'}
}
\description{
Estimating PAMCE withithout regularization
}
