% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AME_fit_basic.R
\name{decompose_AME}
\alias{decompose_AME}
\title{Estimating PAMCE withithout regularization}
\usage{
decompose_AME(outAME, factor_name, level_name, marginal_diff = c("Pop",
  "Exp"))
}
\arguments{
\item{formula}{formula}

\item{data}{data}

\item{pair}{whether we use the paired-conjoint design}

\item{pair_id}{id for paired-conjoint design. required when 'pair = TRUE'}
}
\description{
Estimating PAMCE withithout regularization
}
