% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompose_pAMCE.R
\name{decompose_pAMCE}
\alias{decompose_pAMCE}
\title{Decompose the difference between pAMCEs}
\usage{
decompose_pAMCE(out, effect_name, target_diff)
}
\arguments{
\item{out}{An object of class "pAMCE", a result of a call to 'model_pAMCE'}

\item{target_diff}{Two target profile distributions for which the function compares the pAMCEs}

\item{factor_name}{A factor for which the function visualizes the pAMCEs}

\item{level_name}{A level for which the function visualizes the pAMCEs}
}
\description{
Decompose the difference between pAMCEs
}
