% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_pAMCE.R
\name{design_pAMCE}
\alias{design_pAMCE}
\title{Estimating the population AMCE using a design-based approach}
\usage{
design_pAMCE(formula, factor_name, data, pair = FALSE, pair_id = NULL,
  cross_int = FALSE, cluster_id = NULL, target_dist, target_type)
}
\arguments{
\item{formula}{Formula}

\item{factor_name}{Factors for which the function estimates the pAMCEs. If not specified, the function estimates for all factors.}

\item{data}{Data}

\item{pair}{Whether we use a paired-choice conjoint design}

\item{pair_id}{Unique identifiers for pairs in the paired-choice conjoint design  (optional)}

\item{cross_int}{Include interactions across profiles. Default is FALSE}

\item{cluster_id}{Unique identifiers for computing cluster standard errors (optional)}

\item{target_dist}{Target profile distributions to be used. See Examples for details.}

\item{target_type}{Types of target profile distributions. `marginal`, 'partial_joint', or `target_data`. See Examples for details.}
}
\description{
Estimating the population AMCE using a design-based approach
}
