% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_level.R
\name{estimate_level}
\alias{estimate_level}
\title{Estimating the population AMCE using a model-based approach}
\usage{
estimate_level(object)
}
\arguments{
\item{object}{The output from \code{model_pAMCE} with \code{reg = FALSE}.}
}
\description{
Estimating the population AMCE using a model-based approach
}
