% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorEx.R
\docType{package}
\name{factorEx}
\alias{factorEx}
\alias{factorEx-package}
\title{factorEx}
\description{
This package estimates the PAMCEs with and without regularization
}
\author{
Naoki Egami \email{negami@princeton.edu}
}
