% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_pAMCE.R
\name{model_pAMCE}
\alias{model_pAMCE}
\title{Estimating the population AMCE using a model-based approach}
\usage{
model_pAMCE(formula, formula_three = NULL, data, reg = TRUE, ord_fac,
  pair = FALSE, pair_id = NULL, cross_int = FALSE,
  cluster_id = NULL, target_dist, target_type, difference = FALSE,
  cv_type = "cv.1Std", nfolds = 5, boot = 100, seed = 1234,
  numCores = NULL)
}
\arguments{
\item{formula}{Formula}

\item{formula_three}{Formula for three-way interactions (optional)}

\item{data}{Data}

\item{reg}{TRUE (regularization) or FALSE (no regularization). Default is TRUE}

\item{ord_fac}{Whether we assume each factor is ordered. When not specified, we assume all of them are ordered}

\item{pair}{Whether we use a paired-choice conjoint design}

\item{pair_id}{Unique identifiers for pairs in the paired-choice conjoint design  (optional)}

\item{cross_int}{Include interactions across profiles. Default is FALSE}

\item{cluster_id}{Unique identifiers for computing cluster standard errors (optional)}

\item{target_dist}{Target profile distributions to be used. This argument should be `list`}

\item{target_type}{Types of target profile distributions. `marginal` or `target_data`. See Examples for details}

\item{difference}{Whether we compute the differences between the multiple pAMCEs. Default is FALSE.}

\item{cv_type}{(optimal only when `reg = TRUE``)  `cv.1Std`` (stronger regularization; default) or `cv.min` (weaker regularization).}

\item{nfolds}{Number of cross validation folds. Default is 5.}

\item{boot}{The number of bootstrap samples.}

\item{seed}{Seed for bootstrap.}

\item{numCores}{Number of cores to be used for parallel computing. If not specified, detect the number of available cores internally.}
}
\description{
Estimating the population AMCE using a model-based approach
}
