% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{OnoBurden}
\alias{OnoBurden}
\title{Dataset from Ono and Burden (2018)}
\format{
A list containing the conjoint data, target profile distributions based on (1) marginals, (2) a combination of marginal and partial joint distributions, and (3) the full joint distribution.
}
\usage{
OnoBurden
}
\description{
Dataset from Ono and Burden (2018)
}
\keyword{dataset}
