% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AME_fit_basic.R
\name{plot.cAME}
\alias{plot.cAME}
\title{Estimating PAMCE withithout regularization}
\usage{
\method{plot}{cAME}(came.out, factor_name, main = c("Conditional AMEs"),
  col, pch, mar, legend_pos = "topright", plot_all = TRUE,
  plot_difference = "add", cex = 1)
}
\arguments{
\item{formula}{formula}

\item{data}{data}

\item{pair}{whether we use the paired-conjoint design}

\item{pair_id}{id for paired-conjoint design. required when 'pair = TRUE'}
}
\description{
Estimating PAMCE withithout regularization
}
