% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pAMCE.R
\name{plot.pAMCE}
\alias{plot.pAMCE}
\title{Plotting the estimated population AMCEs}
\usage{
\method{plot}{pAMCE}(x, factor_name, target_dist_name,
  legend_pos = "topright", main = "Estimated population AMCEs", xlim,
  mar = 12, diagnose = FALSE, ...)
}
\arguments{
\item{x}{An object of class "pAMCE", a result of a call to 'model_pAMCE' or 'design_pAMCE'}

\item{factor_name}{Factors for which the function visualizes the pAMCEs}

\item{target_dist_name}{Names of the target profile distributions to be used}

\item{legend_pos}{Position of the legend. Default is 'topright'}

\item{main}{Title of the plot}

\item{xlim}{Range for the x-axis}

\item{mar}{Space on the left side of the plot. Default is 12}

\item{diagnose}{Whether we plot diagnostic checks recommended in de la Cuesta, Egami, and Imai (2019). Default is FALSE}

\item{...}{Other graphical parameters}
}
\description{
Plotting the estimated population AMCEs
}
