% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AME_fit_basic.R
\name{plot_decompose}
\alias{plot_decompose}
\title{Estimating PAMCE withithout regularization}
\usage{
plot_decompose(outAME, factor_name, level_name, type = "decompose",
  marginal_diff, mar = 8)
}
\arguments{
\item{formula}{formula}

\item{data}{data}

\item{pair}{whether we use the paired-conjoint design}

\item{pair_id}{id for paired-conjoint design. required when 'pair = TRUE'}
}
\description{
Estimating PAMCE withithout regularization
}
