% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AME_fit_basic.R
\name{plot_dist}
\alias{plot_dist}
\title{Estimating PAMCE withithout regularization}
\usage{
plot_dist(factor_name, marginal_dist, marginal_type,
  main = "Marginal Distibution", col, pch, mar,
  legend_pos = "topright", cex)
}
\arguments{
\item{formula}{formula}

\item{data}{data}

\item{pair}{whether we use the paired-conjoint design}

\item{pair_id}{id for paired-conjoint design. required when 'pair = TRUE'}
}
\description{
Estimating PAMCE withithout regularization
}
