% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopCon.R
\docType{package}
\name{PopCon}
\alias{PopCon}
\alias{PopCon-package}
\title{PopCon}
\description{
This package estimates the PAMCEs with and without regularization
}
\author{
Naoki Egami \email{negami@princeton.edu}
}
