% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data_collapse.R
\name{prepare_data}
\alias{prepare_data}
\title{Estimating PAMCE withithout regularization}
\usage{
prepare_data(formula, data, marginal_dist, original_level, collapse_level)
}
\arguments{
\item{formula}{formula}

\item{data}{data}

\item{pair}{whether we use the paired-conjoint design}

\item{pair_id}{id for paired-conjoint design. required when 'pair = TRUE'}
}
\description{
Estimating PAMCE withithout regularization
}
