% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_pAMCE.R
\name{summary.pAMCE}
\alias{summary.pAMCE}
\title{Summarizing the estimated population AMCEs}
\usage{
\method{summary}{pAMCE}(object, factor_name, target_dist_name,
  sample = FALSE, ...)
}
\arguments{
\item{object}{An object of class "pAMCE", a result of a call to 'model_pAMCE' or 'design_pAMCE'}

\item{factor_name}{Factors for which the function visualizes the pAMCEs}

\item{target_dist_name}{Names of the target profile distributions to be used}

\item{sample}{Whether to print the sample AMCEs, which is estimated based on the profile distribution used for randomization}

\item{...}{Other parameters}
}
\description{
Summarizing the estimated population AMCEs
}
