% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AME_design.R
\name{weightDesign}
\alias{weightDesign}
\title{Estimating the population AMCE using the design-based estimator}
\usage{
weightDesign(factor_name, target_dist, randomize_dist, sample,
  sim_size = 1000)
}
\arguments{
\item{formula}{Formula}
}
\description{
Estimating the population AMCE using the design-based estimator
}
