% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights_pAMCE.R
\name{weights_pAMCE}
\alias{weights_pAMCE}
\title{Computing weights for estimating the population AMCE using a design-based approach}
\usage{
weights_pAMCE(formula, factor_name, data, pair, pair_id, cross_int,
  target_dist, target_type, partial_joint_name)
}
\arguments{
\item{formula}{Formula}

\item{factor_name}{A factor for which the function computes weights}

\item{data}{Data}

\item{pair}{Whether we use a paired-choice conjoint design}

\item{pair_id}{Unique identifiers for pairs in the paired-choice conjoint design  (optional)}

\item{cross_int}{Include interactions across profiles. Default is FALSE}

\item{target_dist}{Target profile distributions to be used. This argument should be `list`}

\item{target_type}{Types of target profile distributions. `marginal`, `partial_joint` or `target_data`. See Examples for details}

\item{partial_joint_name}{Names of factors representing partial joint distributions. See Examples for details.}
}
\description{
Computing weights for estimating the population AMCE using a design-based approach
}
