% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AME_estimate.R
\name{AME_estimate}
\alias{AME_estimate}
\title{(Internal Function): Estimating PAMCE withithout regularization}
\usage{
AME_estimate(formula, data, pair = FALSE, pair_id = NULL,
  cross_int = TRUE, cluster, marginal_dist, marginal_type,
  joint_dist = NULL, boot, difference = FALSE, formula_three_c)
}
\description{
(Internal Function): Estimating PAMCE withithout regularization
}
