% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AME_estimate_collapse_genlasso.R
\name{AME_estimate_collapse_genlasso}
\alias{AME_estimate_collapse_genlasso}
\title{Estimating PAMCE withithout regularization}
\usage{
AME_estimate_collapse_genlasso(formula, data, ord.fac, pair = FALSE,
  pair_id = NULL, cross_int = TRUE, cluster = NULL, marginal_dist,
  marginal_type, difference = FALSE, cv.type = "cv.1Std", nfolds = 2,
  boot = 100, eps = 1e-04, numCores, seed)
}
\arguments{
\item{formula}{formula}

\item{data}{data}

\item{pair}{whether we use the paired-conjoint design}

\item{pair_id}{id for paired-conjoint design. required when 'pair = TRUE'}
}
\description{
Estimating PAMCE withithout regularization
}
