% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AME_estimate_full.R
\name{AME_estimate_full}
\alias{AME_estimate_full}
\title{Estimating PAMCE withithout regularization}
\usage{
AME_estimate_full(formula, data, type = "genlasso", ord.fac,
  pair = FALSE, pair_id = NULL, cluster = NULL, marginal_dist,
  marginal_type, difference = FALSE, family = "binomial", nway = 1,
  cv.collapse.cost = c(0.1, 0.3, 0.5, 0.7), cv.type = "cv.1Std",
  nfolds = 5, boot = 100, seed = 1234)
}
\arguments{
\item{formula}{formula}

\item{data}{data}

\item{type}{"No-Reg", "gash-anova", or "genlasso"}

\item{ord.fac}{whether we assume each factor is ordered. When not specified, we assume all of them are ordered.}

\item{pair}{whether we use the paired-conjoint design}

\item{pair_id}{id for paired-conjoint design. required when 'pair = TRUE'}

\item{cluster}{id for cluster}

\item{marginal_dist}{marginal distributions}

\item{marginal_type}{names of marginal distributions}

\item{difference}{whether we compute the difference between different estimators}

\item{family}{(only when 'type = gash-anova') when outcomes are binary, "binomial". when outcomes are continuous, "gaussian"}

\item{nway}{(only when `type = gash-anova`) Should be 1 almost always.}

\item{cv.collapse.cost}{(only when `type = gash-anova`) a grid for cross-validation in gash-anova}

\item{cv.type}{(when type = gash-anova or genlasso) `cv.1Std`` (stronger) or `cv.min` (weaker).}

\item{boot}{(when type = gash-anova or genlasso) the number of bootstrap}

\item{seed}{seed for bootstrap}
}
\description{
Estimating PAMCE withithout regularization
}
