% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AME_fit_basic.R
\name{createDist}
\alias{createDist}
\title{Create Profile Distributions from data.frame}
\usage{
createDist(formula, target_data, exp_data, type = "marginal")
}
\arguments{
\item{formula}{formula}

\item{data}{data.frame}

\item{marginal}{Whether we create marginal distributions (TRUE) or the two-dimensional joint distributions}
}
\description{
Create Profile Distributions from data.frame
}
