% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose_pAMCE.R
\name{diagnose_pAMCE}
\alias{diagnose_pAMCE}
\title{Diagnose modeling assumptions for the model-based approach}
\usage{
diagnose_pAMCE(x, factor_name)
}
\arguments{
\item{x}{An object of class "pAMCE", a result of a call to 'model_pAMCE'.}

\item{factor_name}{A factor for which the function diagnoses modeling assumptions.}
}
\description{
Diagnose modeling assumptions for the model-based approach
}
