% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AME_fit_basic.R
\name{plot_AME}
\alias{plot_AME}
\title{Estimating PAMCE withithout regularization}
\usage{
plot_AME(ame.out, factor_name, col, pch, legend_pos = "topright",
  plot_difference = "both", main = "AMEs", xlim, mar, cex = 1,
  plot_type, plot_name)
}
\arguments{
\item{formula}{formula}

\item{data}{data}

\item{pair}{whether we use the paired-conjoint design}

\item{pair_id}{id for paired-conjoint design. required when 'pair = TRUE'}
}
\description{
Estimating PAMCE withithout regularization
}
