% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose_pAMCE.R
\name{plot_diagnose}
\alias{plot_diagnose}
\title{Plotting diagnostic checks}
\usage{
plot_diagnose(x, factor_name, legend_pos = "topright", target_dist_name,
  xlim, mar)
}
\arguments{
\item{x}{An object of class "pAMCE", a result of a call to 'model_pAMCE.'}

\item{factor_name}{A factor for which the function diagnoses modeling assumptions.}

\item{legend_pos}{Position of the legend. Default is 'topright'.}

\item{target_dist_name}{Names of the target profile distributions to be used.}

\item{xlim}{Range for the x-axis.}

\item{mar}{Space on the left side of the plot. Default is 12.}
}
\description{
Plotting diagnostic checks
}
